change: pass if changed object does not exist

See merge request infrastructure/netbox!104

+++ netbox/change/models.py
@@ -659,10 +659,10 @@ class ChangedObject(models.Model):
 
     def apply(self):
         if self.deleted:
-            try:
+            # we don't need to do anything if the object does not exist
+            if not self.changed_object:
+                return
                 self.changed_object.delete()
-            except self.changed_object.__class__.DoesNotExist:
-                pass
         else:
             obj = pickle.loads(self.changed_object_data)
             obj.save(force_insert=True)
